+++ scriptconfig/modal.py
@@ -1,33 +1,58 @@
+"""
+The scriptconfig ModalCLI
+
+This module defines a way to group several smaller scriptconfig CLIs into a
+single parent CLI that chooses between them "modally". E.g. if we define two
+configs: do_foo and do_bar, we use ModalCLI to define a parent program that can run
+one or the other. Let's make this more concrete.
+
+CommandLine:
+    xdoctest -m scriptconfig.modal __doc__:0
+
+Example
+
+    >>> import scriptconfig as scfg
+    >>> #
+    >>> class DoFooCLI(scfg.DataConfig):
+    >>>     __command__ = 'do_foo'
+    >>>     option1 = scfg.Value(None, help='option1')
+    >>>     #
+    >>>     @classmethod
+    >>>     def main(cls, cmdline=1, **kwargs):
+    >>>         self = cls.cli(cmdline=cmdline, data=kwargs)
+    >>>         print('Called Foo with: ' + str(self))
+    >>> #
+    >>> class DoBarCLI(scfg.DataConfig):
+    >>>     __command__ = 'do_bar'
+    >>>     option1 = scfg.Value(None, help='option1')
+    >>>     #
+    >>>     @classmethod
+    >>>     def main(cls, cmdline=1, **kwargs):
+    >>>         self = cls.cli(cmdline=cmdline, data=kwargs)
+    >>>         print('Called Bar with: ' + str(self))
+    >>> #
+    >>> #
+    >>> class MyModalCLI(scfg.ModalCLI):
+    >>>     __version__ = '1.2.3'
+    >>>     foo = DoFooCLI
+    >>>     bar = DoBarCLI
+    >>> #
+    >>> modal = MyModalCLI()
+    >>> MyModalCLI.main(argv=['do_foo'])
+    >>> #MyModalCLI.main(argv=['do-foo'])
+    >>> MyModalCLI.main(argv=['--version'])
+    >>> try:
+    >>>     MyModalCLI.main(argv=['--help'])
+    >>> except SystemExit:
+    >>>     print('prevent system exit due to calling --help')
+"""
+
 import ubelt as ub
+from scriptconfig.util.util_class import class_or_instancemethod
 # from scriptconfig.config import MetaConfig
 
 
 DEFAULT_GROUP = 'commands'
-
-
-class class_or_instancemethod(classmethod):
-    """
-    Allows a method to behave as a class or instance method [SO28237955]_.
-
-    References:
-        .. [SO28237955] https://stackoverflow.com/questions/28237955/same-name-for-classmethod-and-instancemethod
-
-    Example:
-        >>> class X:
-        ...     @class_or_instancemethod
-        ...     def foo(self_or_cls):
-        ...         if isinstance(self_or_cls, type):
-        ...             return f"bound to the class"
-        ...         else:
-        ...             return f"bound to the instance"
-        >>> print(X.foo())
-        bound to the class
-        >>> print(X().foo())
-        bound to the instance
-    """
-    def __get__(self, instance, type_):
-        descr_get = super().__get__ if instance is None else self.__func__.__get__
-        return descr_get(instance, type_)
 
 
 class MetaModalCLI(type):
@@ -182,6 +207,11 @@ class ModalCLI(metaclass=MetaModalCLI):
             self.description = description
 
         self._instance_subconfigs = sub_clis + self.__subconfigs__
+
+        if version is None:
+            version = getattr(self.__class__, '__version__', None)
+        if version is None:
+            version = getattr(self.__class__, 'version', None)
         self.version = version
 
     def __call__(self, cli_cls):
@@ -249,19 +279,8 @@ class ModalCLI(metaclass=MetaModalCLI):
 
             if isinstance(cli_cls, ModalCLI) or issubclass(cli_cls, ModalCLI):
                 # Another modal layer
-                modal = cli_cls
-                from scriptconfig import argparse_ext
-                # FIXME: ModalCLI needs a _parserkw method.
-
-                parserkw.update(
-                    # prog=self._prog,
-                    description=self.description,
-                    # epilog=self._epilog,
-                    # formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-                    # formatter_class=argparse.RawDescriptionHelpFormatter,
-                    formatter_class=argparse_ext.RawDescriptionDefaultsHelpFormatter,
-                    # exit_on_error=False,
-                )
+                modal = cli_cls()
+                parserkw.update(modal._parserkw())
                 parserkw['help'] = parserkw['description'].split('\n')[0]
                 cmdinfo_list.append({
                     'is_modal': True,
@@ -286,15 +305,31 @@ class ModalCLI(metaclass=MetaModalCLI):
                 })
         return cmdinfo_list
 
-    def argparse(self, parser=None, special_options=...):
-
+    def _parserkw(self):
+        """
+        Generate the kwargs for making a new argparse.ArgumentParser
+        """
         from scriptconfig.argparse_ext import RawDescriptionDefaultsHelpFormatter
+        parserkw = dict(
+            description=self.description,
+            formatter_class=RawDescriptionDefaultsHelpFormatter,
+            epilog=getattr(self, '__epilog__', None),
+            prog=getattr(self, '__prog__', None),
+        )
+        if hasattr(self, '__allow_abbrev__'):
+            parserkw['allow_abbrev'] = self.__allow_abbrev__
+        return parserkw
+
+    def argparse(self, parser=None, special_options=...):
+        """
+        Builds a new argparse object for this ModalCLI or extends an existing
+        one with it.
+        """
+
         if parser is None:
             import argparse as argparse_mod
-            parser = argparse_mod.ArgumentParser(
-                description=self.description,
-                formatter_class=RawDescriptionDefaultsHelpFormatter,
-            )
+            parserkw = self._parserkw()
+            parser = argparse_mod.ArgumentParser(**parserkw)
 
         if hasattr(self, 'version') and self.version is not None:
             parser.add_argument('--version', action='store_true',
@@ -323,19 +358,46 @@ class ModalCLI(metaclass=MetaModalCLI):
         # for group, cmdinfos in group_to_cmdinfos.items():
         #     ...
 
+        def fuzzy_cmd_names(n):
+            options = []
+            options.append(n)
+            v1 = n.replace('-', '_')
+            if v1 not in options:
+                options.append(v1)
+            v2 = n.replace('_', '-')
+            if v2 not in options:
+                options.append(v2)
+            main_cmd, *aliases = options
+            return main_cmd, aliases
+
         for cmdinfo in cmdinfo_list:
             # group = cmdinfo['group']
             # Add a new command to subparser_group
+
+            main_cmd, aliases = fuzzy_cmd_names(cmdinfo['cmdname'])
+
+            # TODO: enable alternate hyphen/underscore aliases, but supress
+            # them from the help output. Even better would be to handle
+            # argument completion so they aren't clobbered.
+
+            aliases = []
+            parserkw = cmdinfo['parserkw']
+
+            if 'aliases' in parserkw:
+                parserkw['aliases'] = list(parserkw['aliases']) + list(aliases)
+            else:
+                if aliases:
+                    parserkw['aliases'] = aliases
+
             if cmdinfo['is_modal']:
-                modal_cls = cmdinfo['subconfig']
-                modal_inst = modal_cls()
+                modal_inst = cmdinfo['subconfig']
                 modal_parser = command_subparsers.add_parser(
-                    cmdinfo['cmdname'], **cmdinfo['parserkw'])
+                    main_cmd, **parserkw)
                 modal_parser = modal_inst.argparse(parser=modal_parser)
                 modal_parser.set_defaults(main=cmdinfo['main_func'])
             else:
                 subparser = command_subparsers.add_parser(
-                    cmdinfo['cmdname'], **cmdinfo['parserkw'])
+                    main_cmd, **parserkw)
                 subparser = cmdinfo['subconfig'].argparse(subparser)
                 subparser.set_defaults(main=cmdinfo['main_func'])
         return parser
@@ -343,7 +405,7 @@ class ModalCLI(metaclass=MetaModalCLI):
     build_parser = argparse
 
     @class_or_instancemethod
-    def main(self, argv=None, strict=True):
+    def main(self, argv=None, strict=True, autocomplete='auto'):
         """
         Execute the modal CLI as the main script
         """
@@ -352,13 +414,18 @@ class ModalCLI(metaclass=MetaModalCLI):
 
         parser = self.argparse()
 
-        try:
-            import argcomplete
-            # Need to run: "$(register-python-argcomplete xdev)"
-            # or activate-global-python-argcomplete --dest=-
-            # activate-global-python-argcomplete --dest ~/.bash_completion.d
-            # To enable this.
-        except ImportError:
+        if autocomplete:
+            try:
+                import argcomplete
+                # Need to run: "$(register-python-argcomplete xdev)"
+                # or activate-global-python-argcomplete --dest=-
+                # activate-global-python-argcomplete --dest ~/.bash_completion.d
+                # To enable this.
+            except ImportError:
+                argcomplete = None
+                if autocomplete != 'auto':
+                    raise
+        else:
             argcomplete = None
 
         if argcomplete is not None: